% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_kobocruncher.R
\name{app_kobocruncher}
\alias{app_kobocruncher}
\title{Shiny App - app_kobocruncher}
\usage{
app_kobocruncher(
  maxRequestSize = 50,
  debug = FALSE,
  theme = "IHSN",
  ...,
  shiny.server = FALSE
)
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}

\item{debug}{logical if \code{TRUE}, set shiny-debugging options}

\item{theme}{select stylesheet for the interface. Supported choices are
\itemize{
\item 'yeti'
\item 'flatly'
\item 'journal'
\item 'IHSN'
}}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}

\item{shiny.server}{Setting this parameter to \code{TRUE} will return the app in the form of an
object rather than invoking it. This is useful for deploying \code{app_kobocruncher} via \code{shiny-server}.}
}
\value{
starts the interactive graphical user interface
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{

# app_kobocruncher(theme="flatly")
}
