% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_analyse_indicators}
\alias{f_analyse_indicators}
\title{Analyse indicators}
\usage{
f_analyse_indicators(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
coin Updated coin with analysis tables
}
\description{
Takes a coin and outputs an analysis of indicators. The objective is to look
for any "statistically-problematic" indicators, based on:
}
\details{
\itemize{
\item data availability
\item high proportion of repeated data values
\item outliers as defined by skew and kurtosis
\item collinearity within first aggregation level (category)
\item negative correlation within first aggregation level (category)
}

The output can be viewed in R or as an interactive data frame.
Uses Spearman rank correlation to deal with skewed distributions.
Operates on the Raw data set.

The output is a coin with the results attached. This is so that on export, the
results will also be exported easily.
}
