% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_clean.R
\name{kobo_clean}
\alias{kobo_clean}
\title{Apply cleaning to selected variables within the xlsform}
\usage{
kobo_clean(datalist, dico, indicatoradd = NULL)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}
}
\description{
Basic data cleaning is often required. This functions covers tow use cases:
Use Case 1 - Need to re-categorize the modalities for a select_one or select_multiple
Use case 2-  Need to re-encode one variable based on a cleaned open text question

To address those situation, the kobo_clean function is building, if not set-up,  a cleaning log
for all the variables where the clean cell is not empty within the extended xlsform.

The cleaning log has only 3 columns: variable - old value - new value
This cleaning log can be then easily be adjusted in excel and used in the cleaning workflow when
running the function.

If the cell \code{clean} for the variables refers to the name of another variable that is also included
in the cleaning log, then this second variable will be used to clean the first one from the second one.
This behavior allow to manage the use case 2.

Note that this function can also be used to adjust in order to treat the data and reduce data disclosure risk
See kobo_anonymise
}
\examples{
# dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
# datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )
# 
# kobo_clean(datalist = datalist,
#                    dico = dico,
#                 indicatoradd = indicatoradd   )


 
}
