% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_cruncher.R
\name{kobo_cruncher}
\alias{kobo_cruncher}
\title{Crunch a variable, used in conjunction with the analysis plan}
\usage{
kobo_cruncher(datalist = datalist, dico = dico)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{path to the xlsform file used to colllect the data}
}
\description{
Crunch a variable, used in conjunction with the analysis plan
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

kobo_cruncher(datalist = datalist,
              dico = dico)


}
