% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_data.R
\name{kobo_data}
\alias{kobo_data}
\title{Data loading}
\usage{
kobo_data(datapath)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}
}
\value{
A "datalist" S3 class object (list) formatted to the specifications of "kobocruncher".
}
\description{
Data loading
}
\examples{
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )
# MainFrame
datalist[1]
# Second Frame - based on presence of repeat within the form, aka nested or
# hierarchical data structure, etc... 
datalist[2]  
}
