% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dico.R
\name{kobo_dico}
\alias{kobo_dico}
\title{Prepare Analysis plan}
\usage{
kobo_dico(xlsformpath)
}
\arguments{
\item{xlsformpath}{path to the (extended) xlsform file used to collect the data}
}
\value{
A "kobodico" S3 class object (list) formatted to the specifications of "kobocruncher".
}
\description{
Prepare Analysis plan
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
# Survey
questions <- as.data.frame(dico[[1]])
knitr::kable(utils::head(questions, 10))
# Choices
responses <- as.data.frame(dico[[2]])
knitr::kable(utils::head(responses, 10))
# Settings
metadata <- as.data.frame(dico[[3]])
knitr::kable(utils::head(metadata, 10))
# Report ToC
toc <- as.data.frame(dico[[4]])
knitr::kable(utils::head(toc, 10))
# Indicator
indicator <- as.data.frame(dico[[5]])
knitr::kable(utils::head(indicator, 10))
}
