% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_frame.R
\name{kobo_frame}
\alias{kobo_frame}
\title{get the correct frame for one selected variable - important when having variables within a repeat}
\usage{
kobo_frame(datalist, dico, var)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{variable}
}
\description{
get the correct frame for one selected variable - important when having variables within a repeat
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

data <- kobo_frame(datalist = datalist,
                   dico = dico,
                   var = "members.sex"   )
knitr::kable(utils::head(data,5))
 
}
