% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_indicator.R
\name{kobo_indicator}
\alias{kobo_indicator}
\title{Apply existing documented indicators and add new ones}
\usage{
kobo_indicator(datalist, dico, indicatoradd = NULL)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{indicatoradd}{a list containing all key information to add a calculated indicator within the analysis plan}
}
\description{
The function goes through steps:
1 - load the indicators,
2 - append the one from inidcatoradd if any,
3 - apply the indicator, i.e. do the calculation,
4 - re-save all the working indicator definition within the extended xlsform
5 - bind the new indicators in the dictionary in order to use the kobo_frame() function for further plotting
6 - rebuild the plan if indicators are allocated to chapter, subchapter
}
\examples{
# dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
# datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )
# 
# indicatoradd <- c(  name =  "has.female.members",
#               type = "select_one",
#               label = "Is ",
#               dataframe = "1",
#               calculation = "")
# 
# kobo_indicator(datalist = datalist,
#                    dico = dico,
#                 indicatoradd = indicatoradd   )


 
}
