% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_likert.R
\name{kobo_likert}
\alias{kobo_likert}
\title{Crunch all likert variables according to the analysis plan}
\usage{
kobo_likert(datalist = datalist, datasource = NULL, dico = dico)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{datasource}{name of the data source to display, if set to NULL - then pulls the form_title within the settings of the xlsform}

\item{dico}{path to the xlsform file used to colllect the data}
}
\description{
Crunch all likert variables according to the analysis plan
}
\examples{
dicolikert <- kobo_dico( xlsformpath = system.file("form_likert.xlsx", package = "kobocruncher") )
datalistlikert <- kobo_data(datapath = system.file("data_likert.xlsx", package = "kobocruncher") )

kobo_likert(datalist = datalistlikert,
              dico = dicolikert,
              datasource = "a great survey!")


}
