% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_ridl.R
\name{kobo_ridl}
\alias{kobo_ridl}
\title{Archive all crunching files in RIDL}
\usage{
kobo_ridl(
  ridl,
  datafolder,
  form,
  name_of_this_file,
  republish = "no",
  visibility = "public",
  stage = "explo_initial"
)
}
\arguments{
\item{ridl}{ridl container where the resources should be added}

\item{datafolder}{folder where the data used by the notebook are stored}

\item{form}{names of the file with the analysis plan}

\item{name_of_this_file}{all files are archived based on the name of notebook you created.
The function automatically get the name of the notebook where it is run from,
using
basename(rstudioapi::getSourceEditorContext()$path )}

\item{republish}{"yes" or "no" Used in order to overwrite. default is "no
an existing ridl resource if required}

\item{visibility}{can be  "public" per default or set to private for obscure reasons..}

\item{stage}{allow to document your analysis stage
\itemize{
\item exploration_initial if your crunching is the very initial basic cleaning, relabeling of your data
\item exploration_advance if your crunching
\item interpretation_prez
\item dissemination_story
}}
}
\value{
nothing all analysis files are added as a resources
}
\description{
RIDL is UNHCR instance of a CKAN server. It is designed to keep track and document dataset within an organisation.
It can also be used to conveniently publish your generated report and save the work you did on a notebook
As you have been working on the data, you want to keep track of it and save your work in a place where it can be useful for other people.
}
\details{
The function saves as resource within the  the RIDL container you used to get the data from :
\itemize{
\item the generated report
\item the analysis plan, aka the extended  xlsform used to record relabeling, clean, indicator creation, question grouping, exploration settings
\item the source notebook
}
}
\examples{

### Example used for each template 
# namethisfile = basename(rstudioapi::getSourceEditorContext()$path )  
### template_1_exploration
# if( params$publish == "yes"){
#   kobo_ridl(ridl = params$ridl, 
#                       datafolder = params$datafolder, 
#                       form = params$form,
#                       name =  namethisfile,
#                       republish = params$republish, 
#                       visibility =  params$visibility,
#                       stage = params$stage) }

### template_2_interpretation 

# if( params$publish == "yes"){
#   kobo_ridl(ridl = params$ridl, 
#                       datafolder = params$datafolder, 
#                       form = params$form,
#                       name =  namethisfile,
#                       republish = params$republish,
#                       visibility =  params$visibility,
#                       stage = "interpretation_prez") }

### template_3_dissemination 

# if( params$publish == "yes"){
#  kobo_ridl(ridl = params$ridl, 
#                       datafolder = params$datafolder, 
#                       form = params$form,
#                       name =  namethisfile,
#                       republish = params$republish,
#                       visibility =  params$visibility,
#                       stage = "dissemination_story") }

 
}
