% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_choice.R
\name{label_choice}
\alias{label_choice}
\title{Get the label for a specific choice if available}
\usage{
label_choice(xlsformpath, x)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to cole}

\item{x}{character with the code for the choice}
}
\description{
Get the label for a specific choice if available
}
\examples{
label_choice(xlsformpath =  system.file("sample_xlsform.xlsx", package = "kobocruncher"), 
              x = "profile.country")

}
