% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_choiceset.R
\name{label_choiceset}
\alias{label_choiceset}
\title{Get all the choices labels options for a specific variable}
\usage{
label_choiceset(xlsformpath, x)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to cole}

\item{x}{variable}
}
\description{
Get all the choices labels options for a specific variable
}
\examples{
label_choiceset(xlsformpath =  system.file("sample_xlsform.xlsx", package = "koboloadeR"), 
                x="profile.country")
}
