% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_varhint.R
\name{label_varhint}
\alias{label_varhint}
\title{Get Interpretation hint for a specific variable}
\usage{
label_varhint(dico, x)
}
\arguments{
\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{x}{variable}
}
\description{
Get Interpretation hint for a specific variable
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )

label_varhint(dico = dico, 
              x ="profile.country")
}
