% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_varname.R
\name{label_varname}
\alias{label_varname}
\title{Get the label for a specific variable}
\usage{
label_varname(dico, x)
}
\arguments{
\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{x}{character with the variable name}
}
\description{
Get the label for a specific variable
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )

label_varname(dico = dico, 
              x ="profile.country")

}
