% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_header.R
\name{output_header}
\alias{output_header}
\title{Output Header}
\usage{
output_header(xlsformpath = xlsformpath, var)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to colllect the data}

\item{var}{name of the variable to display}
}
\description{
Output Header
}
\examples{
output_header( xlsformpath =  system.file("sample_xlsform.xlsx", package = "kobocruncher"), 
              var = "profile")
}
