% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plo_correlation.R
\name{plo_correlation}
\alias{plo_correlation}
\title{Plotting Correlation}
\usage{
plo_correlation(
  datapath = datapath,
  xlsformpath = xlsformpath,
  var,
  showcode = FALSE
)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}

\item{xlsformpath}{path to the xlsform file used to cole}

\item{var}{name of the variable to display}

\item{showcode}{display the code}
}
\description{
Plotting Correlation
}
\examples{
plo_correlation(datapath = system.file("data.xlsx", package = "kobocruncher"),
              xlsformpath =  system.file("sample_xlsform.xlsx", package = "kobocruncher"), 
              var = "profile.occupation")
}
