% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_header.R
\name{plot_header}
\alias{plot_header}
\title{Output Header}
\usage{
plot_header(dico = dico, var)
}
\arguments{
\item{dico}{path to the xlsform file used to colllect the data}

\item{var}{name of the variable to display}
}
\value{
text formatted as markdown
}
\description{
Function to add headings within the crunching report -
Headings are by defaults the groups defined in the xlsform -
but can be replaced within the analysis plan by chapter and subchapter
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )

plot_header( dico = dico, 
              var = "profile.profile")

# class(plot_header( dico = dico, 
#               var = "profile.profile"))
# 
dput(plot_header( dico = dico,
              var = "profile.profile"))
# 
message(plot_header( dico = dico, 
               var = "profile.profile"))

cat(plot_header( dico = dico, 
               var = "profile.profile"))

print(plot_header( dico = dico, 
              var = "profile.profile"),
      useSource = FALSE)

}
