% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_one.R
\name{plot_select_one}
\alias{plot_select_one}
\title{Plotting Select one variable}
\usage{
plot_select_one(
  datalist,
  dico,
  var,
  datasource = NULL,
  n = NULL,
  showcode = FALSE
)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{name of the variable to display}

\item{datasource}{name of the data source to display, if set to NULL - then pulls the form_title within the settings of the xlsform}

\item{n}{if not NULL, lumps all levels except for the n most frequent (or least frequent if n < 0) - cf
forcats::fct_lump_n()}

\item{showcode}{display the code}
}
\description{
Note that if the column order is set in the xlsform choice part, the variable will be de factor considered as ordinal and the default ordering will not be done based on frequency
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_select_one(datalist = datalist,
              dico = dico, 
              var = "profile.country",
              showcode = TRUE)
## Exmaple with lumping
plot_select_one(datalist = datalist,
              dico = dico, 
              var = "profile.country",
              n = 1,
              showcode = TRUE)

# plot_select_one(datalist = datalist,
#               dico = dico, 
#               var = "profile.countryerror",
#               showcode = TRUE)
}
