% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_one_cross.R
\name{plot_select_one_cross}
\alias{plot_select_one_cross}
\title{Plotting Select one variable}
\usage{
plot_select_one_cross(
  datapath = datapath,
  xlsformpath = xlsformpath,
  var,
  crosstab,
  showcode = FALSE
)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}

\item{xlsformpath}{path to the xlsform file used to cole}

\item{var}{name of the variable to display}

\item{crosstab}{variable to use for cross tabulation}

\item{showcode}{display the code}
}
\description{
Plotting Select one variable
}
\examples{
plot_select_one_cross(datapath = system.file("data.xlsx", package = "koboloadeR"),
              xlsformpath =  system.file("sample_xlsform.xlsx", package = "koboloadeR"), 
              var = "profile.country",
              crosstab = "profile.occupation"
              )
}
