% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{plot_text}
\alias{plot_text}
\title{Plotting Open Text variables}
\usage{
plot_text(
  datapath = datapath,
  xlsformpath = xlsformpath,
  var,
  showcode = FALSE
)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}

\item{xlsformpath}{path to the xlsform file used to cole}

\item{var}{name of the variable to display}

\item{showcode}{display the code}
}
\description{
Plotting Open Text variables
}
\examples{
plot_text(datapath = system.file("data.xlsx", package = "koboloadeR"),
              xlsformpath =  system.file("sample_xlsform.xlsx", package = "koboloadeR"), 
              var = "profile.occupation")
}
