% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_dissemination.R
\name{template_dissemination}
\alias{template_dissemination}
\title{Report Template 3 for Dissemination and Data Story Telling Template
The last template can be used to take note of the data interpretation session.
It will generate a PDF or an paginated HTML page}
\usage{
template_dissemination(
  back_html = TRUE,
  other_css = NULL,
  number_sections = FALSE,
  ...
)
}
\arguments{
\item{back_html}{Back-cover HTML including contact information}

\item{other_css}{Add extra css}

\item{number_sections}{Number section headings}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown report
}
\description{
Report Template 3 for Dissemination and Data Story Telling Template
The last template can be used to take note of the data interpretation session.
It will generate a PDF or an paginated HTML page
}
