% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_interpretation.R
\name{template_interpretation}
\alias{template_interpretation}
\title{Convert to UNHCR branded Powerpoint presentation}
\usage{
template_interpretation(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{officedown::rpptx_document}}
}
\value{
A powerpoint report
}
\description{
Format for converting from R Markdown to an UNHCR branded Powerpoint presentation
}
