% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_data.R
\name{kobo_data}
\alias{kobo_data}
\title{Data loading}
\usage{
kobo_data(datapath)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}
}
\description{
Data loading
}
\examples{
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "koboloadeR") )
# MainFrame
datalist[1]
# Second Frame
datalist[2]  
}
