% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dico.R
\name{kobo_dico}
\alias{kobo_dico}
\title{Prepare Analysis plan}
\usage{
kobo_dico(xlsformpath)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to collect the data}
}
\description{
Prepare Analysis plan
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
# Survey
questions <- as.data.frame(dico[1])
knitr::kable(head(questions, 10))
# Choices
responses <- as.data.frame(dico[2])
knitr::kable(head(responses, 10))
# Settings
metadata <- as.data.frame(dico[3])
knitr::kable(head(metadata, 10))
}
