% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_frame.R
\name{kobo_frame}
\alias{kobo_frame}
\title{get the correct frame for one selected variable - important when havving variables within a repeat}
\usage{
kobo_frame(datapath, xlsformpath, var)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}

\item{xlsformpath}{path to the xlsform file used to collect the data}

\item{var}{variable}
}
\description{
get the correct frame for one selected variable - important when havving variables within a repeat
}
\examples{
data <- kobo_frame(datapath = system.file("data.xlsx", package = "koboloadeR"),
                   xlsformpath = system.file("sample_xlsform.xlsx", package = "koboloadeR"),
                   var = "members.sex"   )
knitr::kable(head(data,5))
 
}
