% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_prepare_form.R
\name{kobo_prepare_form}
\alias{kobo_prepare_form}
\title{Prepare XLSform and review Analysis}
\usage{
kobo_prepare_form(xlsformpath, xlsformpathout, label_language = "", ridl = "")
}
\arguments{
\item{xlsformpath}{The full path and filename of the xlsform to be accessed (has to be xlsx file)}

\item{xlsformpathout}{The full path and filename of the xlsform to be accessed (has to be xlsx file)}

\item{label_language}{Optional if the form used multiple languages, indicate the language to use to prepare the analysis plan - check first in your original file -default is  ::english (en)}

\item{ridl}{If available, it will prefill the RIDL info through what was already recorded there}
}
\description{
Prepare XLSform by adding instructions for the analysis plan and checking that structure and settings are correct.
This function open the xlsform - extend if required including the excel formatting,
display the analysis plan summary and resave the file at the end.

Once those elements are set up, they will be automatically considered during the automatic crunching phase.
An additional worksheet is also created to document the information required for registration on UNHCR CKAN instance http://ridl.unhcr.org
\enumerate{
\item Configuration of how questions are grouped together in the report:
\itemize{
\item chapter: by default the crunching report is presented according to the group. Once set, this will replace the original grouping. Only variable defined within a chapter will be displayed in the crunching report. By dafault chapter will follow the questions sequence - if chapters start with a number that number will overrule the sequence
\item subchapter: provides a second level of details below the chapter if subchapters start with a number that number will overrule the sequence
}
\item Configuration for data manipulation:
\itemize{
\item clean: define what variable shall be re-categorized during cleaning - a local copy of all levels will be locally saved in order to do the mapping in excel. When the mapping is available, it will be automatically applied to the data. Can be useful to reduce the number of categories.
\item anonymise: define what variables to consider to statistical disclosure risk measurement and subsequent data treatment
}
\item Configuration of specific charts, visualization and analysis:
\itemize{
\item disaggregation: define variable to use for visual cross tabulation - functions with "_cross"
\item correlate: define the variable to use to explore statistical association - works under certain restrictions (i.e. between 2 categorical variables only): kobo_correlate
\item cluster: define variable to generate an unsupervised classification (i.e. hierarchical clustering based on multiple correspondance analysis) kobo_cluster
\item predict: define variable to use to generate predictive model, ie.e the target variable and the predictors.  kobo_predict
\item score: define the different dimensions of a score - and used the score set up for the choice test different aggregation approaches
\item mappoint, mappoly: define the variable to use to generate maps - kobo_map
}
}

In case if those fields do not yet exist, the function will create dummy column for each one.
Also, coloring all rows that have type equal to "begin group", "end group", "begin repeat" or "end repeat" for better legibility
}
\examples{
# kobo_prepare_form(xlsformpath = system.file("form.xlsx", package = "kobocruncher"),
#                   xlsformpathout = "form_with_plan.xlsx")
}
