% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_choiceset.R
\name{label_choiceset}
\alias{label_choiceset}
\title{Get all the choices labels options for a specific variable if available}
\usage{
label_choiceset(dico, x)
}
\arguments{
\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{x}{variable}
}
\description{
This labeling function ia function factory - https://adv-r.hadley.nz/function-factories.html
The output of this function is actually a function
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

data <- kobo_frame(datalist = datalist,
                   dico = dico,
                   var = "profile.country"   )

label_choiceset(dico = dico, 
                x="profile.country")(data$profile.country)

## Test when there's no dictionnary
data$profile.occupation
label_choiceset(dico = dico, 
                x="profile.occupation")(data$profile.occupation)


label_choiceset(dico = dico, 
                x="profile.occupation")(data$profile.occupation)

}
