% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_varhint.R
\name{label_varhint}
\alias{label_varhint}
\title{Get Interpretation hint for a specific variable}
\usage{
label_varhint(xlsformpath, x)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to cole}

\item{x}{variable}
}
\description{
Get Interpretation hint for a specific variable
}
\examples{
label_varhint(xlsformpath = system.file("sample_xlsform.xlsx", package = "koboloadeR"), 
              x ="profile.country")
}
