% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_varname.R
\name{label_varname}
\alias{label_varname}
\title{Get the label for a specific variable}
\usage{
label_varname(xlsformpath, x)
}
\arguments{
\item{xlsformpath}{path to the xlsform file used to collect the data}

\item{x}{character with the variable name}
}
\description{
Get the label for a specific variable
}
\examples{

label_varname(xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher"), 
              x ="profile.country")

}
