% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R, R/plot_select_one_cross.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plotting Correlation}
\usage{
plot_correlation(
  datalist = datalist,
  dico = dico,
  var,
  by_var,
  showcode = FALSE
)

plot_correlation(
  datalist = datalist,
  dico = dico,
  var,
  by_var,
  showcode = FALSE
)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{name of the variable to display}

\item{by_var}{variable to use for cross tabulation}

\item{showcode}{display the code}
}
\description{
blbla

blbla
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_correlation(datalist = datalist,
              dico = dico, 
              var = "profile.occupation")
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_correlation(datalist = datalist,
              dico = dico, 
              var = "profile.occupation")
}
