% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integer.R
\name{plot_integer}
\alias{plot_integer}
\title{Plotting numeric variable}
\usage{
plot_integer(
  datapath = datapath,
  xlsformpath = xlsformpath,
  var,
  showcode = FALSE
)
}
\arguments{
\item{datapath}{path to the file with the data format as extracted from kobo with dot as group separator and xml header}

\item{xlsformpath}{path to the xlsform file used to cole}

\item{var}{name of the variable to display}

\item{showcode}{display the code}
}
\description{
Plotting numeric variable
}
\examples{
plot_integer(datapath = system.file("data.xlsx", package = "kobocruncher"),
              xlsformpath =  system.file("sample_xlsform.xlsx", package = "kobocruncher"), 
              var = "members.age")
}
