% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integer.R, R/plot_select_one.R
\name{plot_integer}
\alias{plot_integer}
\title{Plotting numeric variable}
\usage{
plot_integer(datalist = datalist, dico = dico, var, showcode = FALSE)

plot_integer(datalist = datalist, dico = dico, var, showcode = FALSE)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{name of the variable to display}

\item{showcode}{display the code}
}
\description{
Plotting numeric variable

Plotting numeric variable
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_integer(datalist = datalist,
              dico = dico, 
              var = "members.age",
              showcode = TRUE)
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_integer(datalist = datalist,
              dico = dico, 
              var = "members.age",
              showcode = TRUE)
}
