% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likert.R
\name{plot_likert}
\alias{plot_likert}
\title{Plotting Likert}
\usage{
plot_likert(
  datalist = datalist,
  dico = dico,
  scopei,
  list_namei,
  repeatvari,
  datasource = NULL,
  showcode = FALSE
)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{scopei}{group in which the likert frame are}

\item{list_namei}{name of the likert option list}

\item{repeatvari}{name of the frame within the dataset where to look for the data}

\item{datasource}{name of the data source to display, if set to NULL - then pulls the form_title within the settings of the xlsform}

\item{showcode}{display the code}
}
\description{
Detect if we have more than 3 questions with the same response options within the same questions group
and represent the result using a standard likert plot - build from  https://github.com/jbryer/likert
}
\examples{
dicolikert <- kobo_dico( xlsformpath = system.file("form_likert.xlsx", package = "kobocruncher") )
datalistlikert <- kobo_data(datapath = system.file("data_likert.xlsx", package = "kobocruncher") )

plot_likert(datalist = datalistlikert,
            dico = dicolikert,
            datasource = NULL,
            scopei =  "group_ei8jz33",
            repeatvari =   "main",
            ## getting the list_name and corresponding label
            list_namei = "yk0td68" 
          )
}
