% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_one_cross.R
\name{plot_select_one_cross}
\alias{plot_select_one_cross}
\title{Plotting Select one variable with cross tabulation on a second categorical variable}
\usage{
plot_select_one_cross(
  datalist = datalist,
  dico = dico,
  var,
  by_var,
  showcode = FALSE
)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{name of the variable to display}

\item{by_var}{variable to use for cross tabulation}

\item{showcode}{display the code}
}
\description{
Note that if the column order is set in the xlsform choice part, the variable will be de factor considered as ordinal and the default ordering will not be done based on frequency
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_select_one_cross(datalist = datalist,
              dico = dico, 
              var = "profile.country",
              by_var = "profile.occupation",
              showcode = TRUE
              )
## test if variable are not in the same frame...
plot_select_one_cross(datalist = datalist,
              dico = dico, 
              var = "profile.country",
              by_var = "members.sex",
              showcode = TRUE
              )

}
