% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_one.R, R/plot_text.R
\name{plot_text}
\alias{plot_text}
\title{Plotting Open Text variables}
\usage{
plot_text(datalist = datalist, dico = dico, var, showcode = FALSE)

plot_text(datalist = datalist, dico = dico, var, showcode = FALSE)
}
\arguments{
\item{datalist}{An object of the "datalist" class as defined in kobocruncher}

\item{dico}{An object of the "kobodico" class format as defined in kobocruncher}

\item{var}{name of the variable to display}

\item{showcode}{display the code}
}
\description{
Plotting Open Text variables

Plotting Open Text variables
}
\examples{
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_text(datalist = datalist,
              dico = dico, 
              var = "profile.occupation",
              showcode = TRUE)
dico <- kobo_dico( xlsformpath = system.file("sample_xlsform.xlsx", package = "kobocruncher") )
datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )

plot_text(datalist = datalist,
              dico = dico, 
              var = "profile.occupation",
              showcode = TRUE)
}
