% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_exploration.R
\name{template_exploration}
\alias{template_exploration}
\title{Initial Template for Automatic Data Exploration}
\usage{
template_exploration(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{rmarkdown::html_document}}
}
\value{
An html report
}
\description{
Initial Template for Automatic Data Exploration
}
