% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_exploration.R
\name{template_exploration}
\alias{template_exploration}
\title{Initial Template for Automatic Data Exploration
The first RMD template gives an output in HTML for easy navigation - the left menu provides smooth transition.
It includes a function to automatically run throughout all the survey content. During this stage, data cleaning and new variable creation can be performed through iterations
This report also includes each plot syntax so that they can be easyly pasted for the second report}
\usage{
template_exploration(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{rmarkdown::html_document}}
}
\value{
An html report
}
\description{
Initial Template for Automatic Data Exploration
The first RMD template gives an output in HTML for easy navigation - the left menu provides smooth transition.
It includes a function to automatically run throughout all the survey content. During this stage, data cleaning and new variable creation can be performed through iterations
This report also includes each plot syntax so that they can be easyly pasted for the second report
}
