% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_interpretation.R
\name{template_interpretation}
\alias{template_interpretation}
\title{Second Template to prepare a presentation for the Joint Data Interpretation Session}
\usage{
template_interpretation(...)
}
\arguments{
\item{...}{extra parameters to pass to \code{officedown::rpptx_document}}
}
\value{
A powerpoint report
}
\description{
See a more detailed presentation of that step here: https://www.youtube.com/watch?v=0jE-Y7g88K4&feature=youtu.be&t=2305
}
